requirements.txt
pytest==9.0.3
pytest-playwright==0.7.2
allure-pytest==2.13.5
requests==2.32.3
axe-playwright-python==0.1.7
